+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This changelog documents all notable and breaking changes to the ScurryPy PyPi package.
 
+## [2.3.2] - 19 Sept 2026
+
+* Typing cleanup
+* DiscordError can represent more than just JSON
+
 ## [2.3.1] - 16 Sept 2026
 
 ScurryPy now requires Python 3.12+